app/requirements.txt
mlflow
boto3
psycopg2-binary
psutil==7.0.0
fastapi[standard]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
